Set up for testing (failing test - async function makes it sad)
