+++ src/data/parser-test.csv
@@ -1,0 +1,20 @@
+Section,Description,Category,Category Description,Possible Score,Fred Analysis,Fred Score,,Zana Analysis,Zana Score
+How They Talk About It,Description of how they talk about it,,,,,,,,
+,,"A national, comprehensive mobilization","A national call-to-arms, once-in-a-generation, comprehensive mobilization of all sectors—not just a “climate bill,” a single policy, or a set of policies, but a governing vision and agenda. Recognizes that the Green New Deal is not necessarily a single, omnibus piece of legislation but will, in its entirety, provide a comprehensive, integrated, and systemic response to the crises of climate, inequality, and democracy at hand.",4,"Vice President Fred’s plan lays out a framework for climate action that recognizes the urgency and scale of the climate crisis. The language in the introductory section stresses the need to take immediate action. Fred’s plans recognize the impact the climate crisis has on all aspects of government operation including environment, economy and inequality. He does not, however, go as far as to talk about the Green New Deal as a governing vision or agenda or demonstrate that climate is a lens that will be woven into all action.",3,,"Senator Zana’ plan acknowledges the severity of the crisis, and acknowledges the necessary scope and scale of the solution. The introduction compares the climate crisis to World War II and promises a mobilization on a similar scale to work across all issues of policy to decarbonize our economy and provide a just transition for all.",4
+,,Plans include policies focused on creating jobs,"Candidates’ plans should include policies focused on creating high quality jobs and careers and strengthening people’s ability to live dignified, economically safe lives as central and fundamental pieces of their framing.",4,Vice President Fred’s plan offers strong and consistent framing around high quality job creation.,4,,"Sen. Zana’ plan commits to “ending unemployment by creating 20 million jobs.” This commitment is supported by numerous job-creating projects, including building out clean energy capacity, mass transit infrastructure, and creating a modern Civilian Conversation Corps.",4
+,,Subtotal,,,,7,,,8
+,,,,,,,,,
+How Much They Talk About It,Description of how they talk about it,,,,,,,,
+,,Subtotal,,,,8,,,23
+,,,,,,,,,
+Green New Deal Vision,Description of Green New Deal Vision,,,,,,,,
+1. Emissions Reduction Ambition (20),Description of Emissions Reduction Ambition,,,,,,,,
+,,Mobilizing Resources,"Includes a 10-year mobilization time frame i.e. specific, aggressive actions that will be taken over the course of a decade from commencement of such actions. Investments and policies must be adequate to meet the plan’s stated goals.",4,,1,,,4
+,,Emissions Timeline,"Includes interim targets and a clear road map on how to achieve net-zero emissions* as fast as possible, as well as a society that produces zero or negative emissions well before 2050 so that the United States is doing its fair share to meet the global reductions necessary to prevent 1.5ºC of warming. ",4,,1,,,4
+,,Includes a timeline for:,TODO - this category is particularly weird,9,,,,,
+,,Subtotal,,,,2,,,8
+,,,,,,,,,
+2. Jobs and Economic Security (20),Description of Jobs and Economic Security,,,,,,,,
+,,Job Training,"Includes plans for training and workforce development, and new educational opportunities aimed at preparing all Americans to work in the new careers created by the GND.",2,,1,,,2
+,,Creating High-Quality Jobs,"Includes measures to ensure all jobs are high quality (ex. guaranteed living wages, collective bargaining, local hiring, etc.).",2,,2,,,3
+,,Subtotal,,,,3,,,5
